+++ include/Realm/RLMMongoCollection_Private.h
@@ -1,0 +1,1 @@
+../../Realm/RLMMongoCollection_Private.h